+++ back-end/src/main/java/com/oleksii/leheza/projects/carmarket/dto/view/VehicleGarageDto.java
@@ -17,7 +17,7 @@ public class VehicleGarageDto {
     private String engine;
     private int year;
     private int price;
-    private int mileage;
+    private long mileage;
     private String brandName;
     private String modelName;
     private VehicleStatus status;
